Create switch between local and live datasource
